Extract logic to attach file to a record to a background job
- If the record is not found it will retrun a 404, this is an accepted status code and the request won't be retried by Amazon SNS.
- According to Amazon SNS docs:
  Make sure that your endpoint responds to the HTTP POST message from Amazon SNS with the appropriate status code. The connection will time out in 15 seconds. If your endpoint does not respond before the connection times out or if your endpoint returns a status code outside the range of 200–4xx, Amazon SNS will consider the delivery of the message as a failed attempt.
