+++ app/controllers/mobile_workflow/sns_notifications_controller.rb
@@ -10,26 +10,14 @@ module MobileWorkflow
         when 'SubscriptionConfirmation'
           confirm_subscription ? (head :ok) : (head :bad_request)
         else
-          add_attachment
+          AddAttachmentJob.perform_later(object, object_key, attribute_name)
+          head :ok
         end
+      rescue ActiveRecord::RecordNotFound
+        head :not_found
       end
  
       private
-      def add_attachment
-        begin
-          @object = find_object
-          @object.send("#{attribute_name}=",active_record_blob)
-          if @object.save
-            head :ok
-          else
-            Rails.logger.warn "Error saving object: #{@object} #{object.errors.full_messages}"
-            head :unprocessable_entity
-          end
-        rescue NameError => e
-          Rails.logger.warn "Error attaching object: #{e.message}"
-          head :unprocessable_entity
-        end
-      end
        
       def verify_request_authenticity
         head :unauthorized if raw_post.blank?
@@ -37,22 +25,11 @@ module MobileWorkflow
         #head :unauthorized if raw_post.blank? || !message_verifier.authentic?(raw_post) # Not working
       end
   
-      def active_record_blob
-        s3_object = s3_bucket.object(object_key)
-        checksum_base64 = checksum_base64(object_key, s3_object)
-        ActiveStorage::Blob.create! key: s3_object.key, filename: s3_object.key, byte_size: s3_object.size, checksum: checksum_base64, content_type: s3_object.content_type
-      end
-  
-      def checksum_base64(object_key, s3_object)
-        path = Tempfile.new(object_key).path
-        s3_object.download_file(path)
-        file = File.new(path)
-        Digest::MD5.file(file).base64digest
-      end
-  
-      def find_object
-        object_class_name, object_id = key_identifiers
-        object_class_name.camelcase.constantize.find(object_id.to_i)
+      def object
+        @object ||= begin
+          object_class_name, object_id = key_identifiers
+          object_class_name.camelcase.constantize.find(object_id.to_i)
+        end
       end
   
       def attribute_name
@@ -91,11 +68,7 @@ module MobileWorkflow
         Rails.logger.warn(e.message)
         return false
       end
-  
-      def s3_bucket
-        Aws::S3::Resource.new(region: ENV['AWS_REGION'], access_key_id: ENV['AWS_ACCESS_ID'], secret_access_key: ENV['AWS_SECRET_KEY']).bucket(ENV['AWS_BUCKET_NAME'])
-      end
-  
+
       def sns_client
         Aws::SNS::Client.new(region: ENV['AWS_REGION'], access_key_id: ENV['AWS_ACCESS_ID'], secret_access_key: ENV['AWS_SECRET_KEY'])
       end
